move clause back. was dumb
